backend/requirements.txt
django==5.1.1
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT